organization-name-agent/requirements.txt
agents-api
aio_pika
aiohttp
data-source-api
googleapis-common-protos
grpcio
grpcio-tools
importlib_resources
lz4
phonetics==1.0.5
python-consul2
pyyaml
rapidfuzz
scikit-learn
shiv
tox>=3.20.1
Unidecode==1.2.0